feature/DBCON-298: moves builder class to packaged class
