+++ src/main/java/org/mule/extension/db/internal/domain/connection/oracle/OracleConnectionParameters.java
@@ -13,6 +13,7 @@ import static org.mule.runtime.extension.api.error.MuleErrors.CONNECTIVITY;
 
 import org.mule.extension.db.internal.domain.connection.BaseDbConnectionParameters;
 import org.mule.extension.db.internal.domain.connection.DataSourceConfig;
+import org.mule.extension.db.internal.domain.connection.oracle.util.OracleTNSEntryBuilder;
 import org.mule.runtime.api.connection.ConnectionException;
 import org.mule.runtime.api.tls.TlsContextFactory;
 import org.mule.runtime.extension.api.annotation.Expression;
@@ -100,7 +101,7 @@ public class OracleConnectionParameters extends BaseDbConnectionParameters imple
     checkInstanceAndServiceName();
 
     StringBuilder buf = new StringBuilder(JDBC_URL_PREFIX);
-    return tlsContextFactory != null ? generateSecureUrl(buf) : generateBasicUrl(buf);
+    return tlsContextFactory != null ? generateSecureUrl() : generateBasicUrl(buf);
   }
 
   @Override
@@ -140,14 +141,14 @@ public class OracleConnectionParameters extends BaseDbConnectionParameters imple
     return buf.toString();
   }
 
-  private String generateSecureUrl(StringBuilder buf) {
+  private String generateSecureUrl() {
     return new OracleTNSEntryBuilder()
         .withProtocol("TCPS")
         .withHost(host)
         .withPort(port)
         .withInstanceName(instance)
         .withServiceName(serviceName)
-        .build(buf);
+        .build();
   }
 
   private void checkInstanceAndServiceName() {
@@ -157,69 +158,6 @@ public class OracleConnectionParameters extends BaseDbConnectionParameters imple
 
       throw new ModuleException(errorMessage, CONNECTIVITY, new ConnectionException(errorMessage));
     }
-  }
-
-  /**
-   *
-   * Builder class to facilitate the creation of Oracle's URLs that use TNS Entries
-   *
-   */
-  public static final class OracleTNSEntryBuilder {
-
-    private String protocol;
-    private String host;
-    private Integer port;
-    private String instanceName;
-    private String serviceName;
-
-    private OracleTNSEntryBuilder() {}
-
-    public OracleTNSEntryBuilder withProtocol(String protocol) {
-      this.protocol = protocol;
-      return this;
-    }
-
-    public OracleTNSEntryBuilder withHost(String host) {
-      this.host = host;
-      return this;
-    }
-
-    public OracleTNSEntryBuilder withPort(Integer port) {
-      this.port = port;
-      return this;
-    }
-
-    public OracleTNSEntryBuilder withInstanceName(String instanceName) {
-      this.instanceName = instanceName;
-      return this;
-    }
-
-    public OracleTNSEntryBuilder withServiceName(String serviceName) {
-      this.serviceName = serviceName;
-      return this;
-    }
-
-    public String build(StringBuilder buf) {
-      buf.append("(DESCRIPTION=");
-
-      buf.append("(ADDRESS=");
-      buf.append("(PROTOCOL=").append(protocol).append(")");
-      buf.append("(PORT=").append(port).append(")");
-      buf.append("(HOST=").append(host).append(")");
-      buf.append(")");
-
-      buf.append("(CONNECT_DATA=");
-      if (instanceName != null) {
-        buf.append("(INSTANCE_NAME=").append(instanceName).append(")");
-      }
-      if (serviceName != null) {
-        buf.append("(SERVICE_NAME=").append(serviceName).append(")");
-      }
-      buf.append(")");
-
-      return buf.append(")").toString();
-    }
-
   }
 
 }
